% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aplica_funcio.R
\name{aplica_funcio}
\alias{aplica_funcio}
\title{selecciona i aplica la funció d'anàlisi a executar.}
\usage{
aplica_funcio(param)
}
\arguments{
\item{param}{vector de paràmetres per la funció. Creat al pas anterior}
}
\value{
vector caràcter amb el codi a executar, amb els noms de les variables a utilitzar
}
\description{
segona peça de la fàbrica de chunks. Tria la funció a aplicar segons si les variables dependents són qualitatives o quantitatives.
}
\examples{
utils::data(package = "palmerpenguins", "penguins")
param <- expand.grid(
  dades = "penguins",
  VD    = c("bill_length_mm", "bill_depth_mm"),
  VI    = c("island"),
  pes   = NULL) \%>\%
  lapply(., function(x){as.list(as.character(x))})
aplica_funcio(param)
}
