% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunkfactory_display.R
\name{chunkfactory_display}
\alias{chunkfactory_display}
\title{chunkfactory_display}
\usage{
chunkfactory_display(myfunc, display_list, title_level = 1)
}
\arguments{
\item{myfunc}{funció que es vol aplicar iterativament sobre la llista de resultats. Veure exemple de funció per display en chunkfactory:::display_reslist()}

\item{display_list}{named list of results to display}

\item{title_level}{nivell dels títols que es desitja per la organització del document}
}
\value{
markup que s'interpreta via knitr::knit_child()
}
\description{
chunkfactory_display
}
\examples{
\dontrun{
chunkfactory_display(myres)
}
}
