% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computa_residusAS.R
\name{computa_residusAS}
\alias{computa_residusAS}
\title{Computes weighted crosstabs, including adjusted standardized residuals.}
\usage{
computa_residusAS(data, VI, VD, pes)
}
\arguments{
\item{data}{dataset}

\item{VI}{Idependent variable (rows), as character string}

\item{VD}{Dependent variable (columns), as character string}

\item{pes}{Weights variable, as character string}
}
\value{
tibble with 2 columns: key1 és la clau primària, una combinació de les categories de les VI & VD, ASres és el valor dels residus ajustats i estandarditzats, segons es calculen a chisquare::chisquare() "stand.res[i,j]/sqrt((1−sr[i]/n)∗(1−sc[j]/n)). Els resultats d'aquesta funció estan pensats per fer servir en combinació amb "
}
\description{
Computes weighted crosstabs, including adjusted standardized residuals.
}
\examples{
data(package = "palmerpenguins", "penguins")
penguins$pes <- 1
computa_residusAS(data = penguins, VI = "species", VD = "sex", pes = "pes")
}
