% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computa_residusAS.R
\name{computa_residusAS_nse}
\alias{computa_residusAS_nse}
\title{Computa residus ajustats i estandarditzat per taules de contingència, partint d'una taula ponderada}
\usage{
computa_residusAS_nse(data, VI, VD, pes)
}
\arguments{
\item{data}{Conjunt de dades}

\item{VI}{Variable independent en NSE (a files)}

\item{VD}{Variable dependent  en NSE (a columnes)}

\item{pes}{Variable amb els pesos en NSE}
}
\value{
tibble amb dues columnes: key1 és la clau primària, una combinació de les categories de les VI & VD, ASres és el valor dels residus ajustats i estandarditzats, segons es calculen a chisquare::chisquare() "stand.res[i,j]/sqrt((1−sr[i]/n)∗(1−sc[j]/n)). Els resultats d'aquesta funció estan pensats per fer servir en combinació amb "
}
\description{
Computa residus ajustats i estandarditzat per taules de contingència, partint d'una taula ponderada
}
