% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_reslist.R
\name{display_reslist}
\alias{display_reslist}
\title{display a list of results in a given order o layout in a markdown document. Works in coordination with chunkfactory_display}
\usage{
display_reslist(myset, mylayout = "layout1")
}
\arguments{
\item{myset}{named list containing set of results to be displayed}

\item{mylayout}{definition of the layout or order that the results should be displayed in}
}
\value{
block of markdown to be interpreted by knitr::knit_child
}
\description{
display a list of results in a given order o layout in a markdown document. Works in coordination with chunkfactory_display
}
\examples{
\dontrun{
display_reslist(myset)
}
}
