% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabrica_chunks_myfunc.R
\name{.chunkify}
\alias{.chunkify}
\title{make a text string that can be parsed by knitr, containing a chunk for each output in the list}
\usage{
.chunkify(reslist, sections)
}
\arguments{
\item{reslist}{list of results as produced by .reslistify}

\item{sections}{text string defining the kind of separation of the sections}
}
\value{
a text string
}
\description{
make a text string that can be parsed by knitr, containing a chunk for each output in the list
}
