% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabrica_chunks_myfunc.R
\name{.reslistify}
\alias{.reslistify}
\title{make reslist structure for fabrica_chunks_myfunc}
\usage{
.reslistify(myfunc, param_list)
}
\arguments{
\item{myfunc}{custom function}

\item{param_list}{named list of parameters}
}
\value{
a list
}
\description{
make reslist structure for fabrica_chunks_myfunc
}
