% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escriu_codi.R
\name{escriu_codi}
\alias{escriu_codi}
\title{function that writes markdown and code}
\usage{
escriu_codi(reslist_mytab, title_level = 3)
}
\arguments{
\item{reslist_mytab}{list of results produced by code factories in previous steps}

\item{title_level}{level of the titles for the sections. Defaults to Header 3}
}
\value{
character vector containing the markup and code later interpreted by knitr::knit_child(text = unlist(src))
}
\description{
function that writes markdown and code
}
\examples{
utils::data(package = "palmerpenguins", "penguins")
# define function parameters
param <- fes_param_list(c("species"), c("island"), "penguins")
# extract results into a list
reslist_mytab <- aplica_funcio(param)
# extract code with results
src <- escriu_codi(reslist_mytab)
}
