% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraccio_RAE.R
\name{extraccio_RAE}
\alias{extraccio_RAE}
\title{Compute adjusted standardized residuals (ASR).}
\usage{
extraccio_RAE(mytab)
}
\arguments{
\item{mytab}{crosstab with absolute frequencies, as produced by table(), or questionr::wtd.table()}
}
\value{
table containing a primary key and AER.
}
\description{
he extret el codi de chisquare::chisquare(), que te altament verbose output, per obtenir només la part de residus ajustats estandarditzats que volia.
}
\examples{
data(package = "palmerpenguins", "penguins")
mytab <- table(penguins$island, penguins$sex)
computa.ASR(mytab)
}
