% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabrica_chunks.R
\name{fabrica_chunks}
\alias{fabrica_chunks}
\title{fabrica els chunks, amb els seus títols, chunks i codi}
\usage{
fabrica_chunks(vd, vi, d, w)
}
\arguments{
\item{vd}{vector de noms de les variables dependents}

\item{vi}{vector de noms de les variables independents}

\item{d}{nom de les dades, com a cadena de caràcter}

\item{w}{nom de la variable de ponderació}
}
\value{
markdown per generar seccions (sota títol 3), chunks i codi que serà posteriorment renderitzat en un rmd.
}
\description{
fabrica els chunks, amb els seus títols, chunks i codi
}
\examples{
utils::data(package = "palmerpenguins", "penguins")
fabrica_chunks(c("sex", "species"), c("island"), "penguins")
}
