% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabrica_chunks_myfunc.R
\name{fabrica_chunks_myfunc}
\alias{fabrica_chunks_myfunc}
\title{Turn any custom function into a chunk factory.}
\usage{
fabrica_chunks_myfunc(myfunc, param_list, sections = "3")
}
\arguments{
\item{myfunc}{custom function you'd like to execute. Data arguments need to be treated as character strings. See example.}

\item{param_list}{list of parameters needed by the function. Need to be specified with the same names, in the same order as the function arguments.}

\item{sections}{level of the titles for the sections. Defaults to Header 3}
}
\value{
a character string containing the markdown generating the document sections  (under a title 3), the chunks and the code to be rendered by the .rmd document.
}
\description{
Turn any custom function into a chunk factory.
}
\examples{
utils::data(package = "palmerpenguins", "penguins")
myfunc <- function(data_name, v1){
  data <- get(data_name)
  res1 <- table(data[[v1]])
  res2 <- ggplot2::ggplot(data) + ggplot2::geom_bar(ggplot2::aes(x = .data[[v1]]))
  res3 <- prop.table(res1)
  return(list(res1, res2, res3))
}
myfunc_params <- list(data_name = "penguins", v1 = c("species", "island", "sex"))
fabrica_chunks_myfunc(myfunc, myfunc_params, sections = "1")
}
