% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fes_param_list.R
\name{fes_param_list}
\alias{fes_param_list}
\title{Define a list of parameters for descriptive statistics functions, in the order needed by the followong functions, notably purrr::pmap.}
\usage{
fes_param_list(d, vi, vd, w)
}
\arguments{
\item{d}{name of the dataset}

\item{vi}{name of the independent variable}

\item{vd}{name of the dependent variable}

\item{w}{name of the weighting variable}
}
\value{
list with all the combinations of the different elements
}
\description{
Define a list of parameters for descriptive statistics functions, in the order needed by the followong functions, notably purrr::pmap.
}
\examples{
\dontrun{
fes_param_list("penguins", c("species"), c("bill_length_mm", "bill_depth_mm"), w = "pes")
}
}
