% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fes_taula_nse.R
\name{fes_taula_nse}
\alias{fes_taula_nse}
\title{Fes una taula de contingència ponderada, fent servir NSE}
\usage{
fes_taula_nse(dades, VD, VI, pes)
}
\arguments{
\item{dades}{dades}

\item{VD}{variable dependent (NSE)}

\item{VI}{variable independent (NSE)}

\item{pes}{pes (NSE)}
}
\value{
una tibble amb els recomptes (n), els recomptes ponderats (N), els totals de VI (TT), els percentatges (PP)
}
\description{
Fes una taula de contingència ponderada, fent servir NSE
}
\examples{
data(package = "palmerpenguins", "penguins")
penguins$pes <- 1
fes_taula_nse(penguins, island, sex, pes)
}
