% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycrosstab.R
\name{mycrosstab}
\alias{mycrosstab}
\title{taula de contingència a mida}
\usage{
mycrosstab(dades, VD, VI, pes)
}
\arguments{
\item{dades}{dades}

\item{VD}{variable dependent (NSE)}

\item{VI}{variable independent (NSE)}

\item{pes}{pes (NSE)}
}
\value{
una tibble amb els recomptes (n), els recomptes ponderats (N), els totals de VI (TT), els percentatges (PP) i els residus ajustats estandarditzats
}
\description{
Calcula una taula de contingència amb residus ajustats estandarditzats, fent servir NSE. Es donen només els % sobre els marginals de la VI.
}
\examples{
data(package = "palmerpenguins", "penguins")
penguins$pes <- 1
mycrosstab(penguins, island, sex, pes)
}
