% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mytab_txt.R
\name{mytab_txt}
\alias{mytab_txt}
\title{Code factory for the analysis of categorical variables.}
\usage{
mytab_txt(dades, VI, VD, pes)
}
\arguments{
\item{dades}{character string with the name of the dataset.}

\item{VI}{character string with the names of the independent variables}

\item{VD}{character string with the names of the dependent variables}

\item{pes}{character string with the names of the weighting variables.}
}
\value{
list with the code to execute a crosstab and a stacked bar chart.
}
\description{
Code factory for the analysis of categorical variables.
}
\examples{
data(package = "palmerpenguins", "penguins")
penguins$pes <- 1
mytab_txt("penguins", "island", "species", "pes")
}
