% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_txt.R
\name{quant_txt}
\alias{quant_txt}
\title{fàbrica de codi per analitzar variables numèriques}
\usage{
quant_txt(dades, VD, VI, pes)
}
\arguments{
\item{dades}{nom del conjunt de dades a analitzar, com a cadena de caràcters}

\item{VD}{nom, o conjunt de noms, de les variables dependents}

\item{VI}{nom, o conjunt de noms, de les variables independents}

\item{pes}{nom de la variable de ponderació. Si no existeix, s'assigna 1}
}
\value{
llista amb el codi per executar la taula d'estadístics escriptius i el diagrama de caixes
}
\description{
fàbrica de codi per analitzar variables numèriques
}
\examples{
data(package = "palmerpenguins", "penguins")
quant_txt("penguins", "island", VI = c("body_mass_g", "bill_length_mm"))
}
