% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aplica_funcio.R
\name{aplica_funcio}
\alias{aplica_funcio}
\title{select and execute analysis function for bivariate descriptive analysis}
\usage{
aplica_funcio(param)
}
\arguments{
\item{param}{vector of parametres for the function.}
}
\value{
character vector caracter with the code to be executed.
}
\description{
second piece in the chunk factory. Select the function according to the class of the dependent variable.
}
\examples{
data(package = "palmerpenguins", "penguins")
penguins$pes <- 1
param <- expand.grid(
  dades = "penguins",
  VI    = c("island"),
  VD    = c("bill_length_mm", "bill_depth_mm"),
  pes   = "pes") \%>\%
  lapply(., function(x){as.list(as.character(x))})
chunkfactory:::aplica_funcio(param)
}
