% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escriu_codi.R
\name{escriu_codi}
\alias{escriu_codi}
\title{funció que escriu codi dins d'un chunk, i amb títols 3 per cada secció}
\usage{
escriu_codi(reslist_mytab)
}
\arguments{
\item{reslist_mytab}{llista de resultats d'aplicar la fàbrica de codi, a les passes anteriors}
}
\value{
vector caràcter amb el markdown i el codi que han d'interpretar-se posteriorment per  knitr::knit_child(text = unlist(src))
}
\description{
funció que escriu codi dins d'un chunk, i amb títols 3 per cada secció
}
\examples{
utils::data(package = "palmerpenguins", "penguins")
# defineix els paràmetres per alimentar a la funció d'anàlisi
param <- fes_param_list(c("species"), c("island"), "penguins")
# extreu resultats en llista
reslist_mytab <- aplica_funcio(param)
# extreu el codi amb resultats
src <- escriu_codi(reslist_mytab)
}
