% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabrica_chunks.R
\name{fabrica_chunks}
\alias{fabrica_chunks}
\title{fabricate chunks, with section titles, chunks delimitations, and code.}
\usage{
fabrica_chunks(d, vi, vd, w)
}
\arguments{
\item{d}{character vector with name of dataset}

\item{vi}{character vector with name of independent variable}

\item{vd}{character vector with name of dependent variable}

\item{w}{character vector with name of weighting variable}
}
\value{
markdown per generar seccions (sota títol 3), chunks i codi que serà posteriorment renderitzat en un rmd.
}
\description{
fabricate chunks, with section titles, chunks delimitations, and code.
}
\examples{
utils::data(package = "palmerpenguins", "penguins")
penguins$pes <- 1
fabrica_chunks("penguins", c("sex", "species"), c("island"), "pes")
}
