% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fes_param_list.R
\name{fes_param_list}
\alias{fes_param_list}
\title{Defineix paràmetres per l'anàlisi}
\usage{
fes_param_list(vd, vi, d, w)
}
\arguments{
\item{vd}{nom o vector de noms de les variables dependents}

\item{vi}{nom o vector de noms de les variables independents}

\item{d}{nom del conjunt de dades}

\item{w}{nom de la variable de ponderació}
}
\value{
llista amb els noms i l'ordre dels arguments tal i com ho requereix purrr::pmap i les funcions d'anàlisi
}
\description{
Defineix paràmetres per l'anàlisi
}
\examples{
fes_param_list(c("bill_length_mm", "bill_depth_mm"), c("species"), "penguins")
}
