% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycrosstab.R
\name{mycrosstab}
\alias{mycrosstab}
\title{tailored cross tabulation}
\usage{
mycrosstab(dades, VI, VD, pes)
}
\arguments{
\item{dades}{dataset}

\item{VI}{name of the independent variable}

\item{VD}{name of the dependent variable}

\item{pes}{name of the weighting variable}
}
\value{
a tibble including counts (n), weighted counts (N), marginals of the independent variable (TT), percentages within independent variable groups (PP = N/TT*100), and adjusted standardized residuals.
}
\description{
Compute a crosstaband add adjusted standardized residuals. Percentages are only computed using independent variable marginals.
}
\examples{
data(package = "palmerpenguins", "penguins")
penguins$pes <- 1
mycrosstab(penguins, "island", "sex", "pes")
}
