% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mytab_txt.R
\name{mytab_txt}
\alias{mytab_txt}
\title{Fàbrica de codi per analitzar variables categòriques}
\usage{
mytab_txt(dades, VD, VI, pes)
}
\arguments{
\item{dades}{nom del conjunt de dades a analitzar, com a cadena de caràcters}

\item{VD}{nom, o conjunt de noms, de les variables dependents}

\item{VI}{nom, o conjunt de noms, de les variables independents}

\item{pes}{nom de la variable de ponderació. Si no existeix, s'assigna 1}
}
\value{
llista amb el codi per executar la taula de freqüències i el diagrama de barres
}
\description{
Fàbrica de codi per analitzar variables categòriques
}
\examples{
data(package = "palmerpenguins", "penguins")
mytab_txt("penguins", "island", "species")
}
