% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_txt.R
\name{quant_txt}
\alias{quant_txt}
\title{code factory to analyse numerical variables.}
\usage{
quant_txt(dades, VI, VD, pes)
}
\arguments{
\item{dades}{character vector with the name of the dataset}

\item{VI}{character vector with the name of the independent variable}

\item{VD}{character vector with the name of the dependent variable}

\item{pes}{character vector with the name of the weighting variable}
}
\value{
list with the code to execute a table of descriptive statistics and grouped boxplots.
}
\description{
code factory to analyse numerical variables.
}
\examples{
data(package = "palmerpenguins", "penguins")
penguins$pes <- 1
chunkfactory:::quant_txt("penguins", "island", VD = c("body_mass_g", "bill_length_mm"), "pes")
}
